+++ spec/mysql_helper.rb
@@ -6,13 +6,7 @@ class MysqlHelper
   MYSQL_USER    = ENV.fetch('ARMG_TEST_MYSQL_USER', 'root')
   MYSQL_DB      = ENV.fetch('ARMG_TEST_MYSQL_DB', 'armg_test')
   MYSQL_ENGINE  = ENV.fetch('ARMG_TEST_MYSQL_ENGINE', 'MyISAM')
-  TABLE_OPTIONS = if ActiveRecord.gem_version < Gem::Version.new('6.1.0')
-                    "ENGINE=#{MYSQL_ENGINE} DEFAULT CHARSET=utf8"
-                  elsif MYSQL_ENGINE == 'InnoDB'
-                    nil
-                  else
-                    "ENGINE=#{MYSQL_ENGINE}"
-                  end
+  TABLE_OPTIONS = "ENGINE=#{MYSQL_ENGINE} DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin"
 
   def initialize
     @mysql = Mysql2::Client.new(
@@ -41,9 +35,9 @@ class MysqlHelper
       t.index ['location'], name: 'idx_location', type: :spatial
     end
 
-    ActiveRecord::Base.connection.execute("INSERT INTO geoms (id, location) VALUES (1, GeomFromText('POINT(1 1)', 1245))")
-    ActiveRecord::Base.connection.execute("INSERT INTO geoms (id, location) VALUES (2, GeomFromText('LINESTRING(0 0,1 1,2 2)'))")
-    ActiveRecord::Base.connection.execute("INSERT INTO geoms (id, location) VALUES (3, GeomFromText('POLYGON((0 0,10 0,10 10,0 10,0 0),(5 5,7 5,7 7,5 7, 5 5))', 5678))")
+    ActiveRecord::Base.connection.execute("INSERT INTO geoms (id, location) VALUES (1, ST_GeomFromText('POINT(1 1)', 3857))")
+    ActiveRecord::Base.connection.execute("INSERT INTO geoms (id, location) VALUES (2, ST_GeomFromText('LINESTRING(0 0,1 1,2 2)'))")
+    ActiveRecord::Base.connection.execute("INSERT INTO geoms (id, location) VALUES (3, ST_GeomFromText('POLYGON((0 0,10 0,10 10,0 10,0 0),(5 5,7 5,7 7,5 7, 5 5))', 3857))")
   end
 
   private