+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Support MySQL 8.0 (with limitation) [#18](https://github.com/cookpad/armg/pull/18)
+
 ### Changed
 
 ### Deprecated